Rename bulk export task creation module to bulk export task creator and update references
