+++ java/bulk-export/bulk-export-task-creator/src/main/java/sleeper/bulkexport/taskcreator/RunLeafPartitionBulkExportTasks.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.bulkexport.taskcreation;
+package sleeper.bulkexport.taskcreator;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
